+++ the_forge/pdf_cache.py
@@ -6,7 +6,7 @@ from django.core.serializers.json import DjangoJSONEncoder
 from django.forms.models import model_to_dict
 
 
-CACHE_PREFIX = 'forge_pdf:v7'
+CACHE_PREFIX = 'forge_pdf:v8'
 PDF_CACHE_TTL = 60 * 60 * 24
 PDF_CACHE_MAX_BYTES = 25 * 1024 * 1024